Refactor SnapshotTagValuePush
`SnapshotTagValuePushBase` is a new base class that contains the majority of logic that was previously in `SnapshotTagValuePush`. `SnapshotTagValueManager` is now derived from the new base class. This allows `SnapshotTagValuePush` to remove values from its cache when there are no subscribers remaining for a tag, whereas `SnapshotTagValueManager` can keep cached values regardless of how many subscribers a tag has.
